refactor: 将数据库操作集中到 OsuPlayerDbContext 中
将 AppDbOperator 和 BeatmapDbOperator 中的数据库逻辑迁移至 OsuPlayerDbContext 的实例方法，使 AppDbOperator 成为 DbContext 的工厂与代理，BeatmapDbOperator 的扩展方法转而调用 DbContext 实例。此举消除了重复的数据库访问代码，统一了事务与异常处理逻辑。旧的 GetDapperConnection 方法标记为过时，引导调用方通过 DbContext 的 Database.GetDbConnection() 共享生命周期。
